Add CSRF tokens and secure some functions

+++ deleteMessage.php
@@ -15,26 +15,26 @@
 
    session_start();
    include("checkUserSession.php");
+   include("functions.php");
    include("databaseConnection.php");
    $messageId = $_GET['messageId'];
 
    // Get the message receiver id	
-   $sql = "SELECT message_receiver_id FROM messages WHERE message_id = $messageId";
-   global $file_db;
-   $result = $file_db->query($sql);
-   $result->setFetchMode(PDO::FETCH_ASSOC);
-   $result = $result->fetch();
+   $result = getReceiverId($messageId);
+   
    $messageReceiverId = $result['message_receiver_id'];
 
+
    // Check if user session exists and if the user has the right to delete the message
-   if ($_SESSION['userId'] == $messageReceiverId) {
+   if ($_SESSION['userId'] == $messageReceiverId && $_GET['CSRFToken'] == $_SESSION["CSRFtoken"]) {
       $sql = "DELETE FROM messages where message_id = $messageId";
       $result = $file_db->query($sql);
 
       header('Location: http://localhost/html/messages.php?result=deleted');
       exit();
    } else {
-      header('Location: http://localhost/html/messages.php?result=notdeleted');
-      exit();
+	   echo $_SESSION["CSRFtoken"];
+      //header('Location: http://localhost/html/messages.php?result=notdeleted');
+      //exit();
    }
 ?>
